Extract storage and battle runner modules
